Shop controllers update

+++ src/main/java/com/kalemz/the_shopper/shop/ShopRepository.java
@@ -2,9 +2,5 @@ package com.kalemz.the_shopper.shop;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.List;
-
-public interface ShopRepository extends JpaRepository<Shop, Long> {
-
-
+public interface ShopRepository extends JpaRepository<Shop, Integer> {
 }
